backend/requirements.txt
# MedFirst Backend Dependencies
fastapi==0.109.2
uvicorn[standard]==0.27.1
python-dotenv==1.0.1
google-genai==1.0.0
google-cloud-speech==2.24.1
google-cloud-firestore==2.14.0
websockets>=13.0
python-multipart==0.0.9
aiofiles==23.2.1
pydantic==2.6.1
